[WIP] new snapshotter: zfs
Usage:
- set snapshotter to zfs in config.toml
- mount a zfs on /var/lib/containerd/snapshot/zfs

I confirmed `dist pull` & `ctr run` works with Ubuntu 17.04 kernel-mode
ZFS.
Should work on other ZFS implementations as well.

Note that BasicTest still fails because Remove() is not fully implemented
(Also, you need to set `-test.zpool=your-zpool` for running `zfs.test`)

+++ linux/runtime.go
@@ -111,6 +111,10 @@ func (r *Runtime) Create(ctx context.Context, id string, opts containerd.CreateO
 			Options: m.Options,
 		})
 	}
+	log.G(ctx).Debugf("shim.CreateRequest.Rootfs=%+v", sopts.Rootfs)
+	if len(sopts.Rootfs) == 0 {
+		return nil, fmt.Errorf("rootfs empty for id %s", id)
+	}
 	if _, err := s.Create(ctx, sopts); err != nil {
 		os.RemoveAll(path)
 		return nil, err
